break out renderConfig.tsx into helpers and fix test failures
